+++ NEWS.md
@@ -1,3 +1,14 @@
+# SRJPEdata 1.0.0
+
+This release contains:
+- RST (including mark recapture trials) data through the 2025/2026 monitoring season (2025/2026 is still considered preliminary). 
+- Adult estimates through 2025 (aligns with 2025/2026 juvenile data), except for Yuba River which has data through 2024
+- Updated flow and water temperature data through May 2026
+
+This dataset will be used to develop a manuscript for the BTSPAS-X model.
+
+Patches are planned for May-June 2026 that involve improvements to water temperature data processing and package documentation. The 2025/2026 RST data is expected to be finalized by July 2026 and released as a minor update to this version.
+
 # SRJPEdata 0.0.2
 
 Updates included in this patch: